services/projects/pyproject.toml
[tool.poetry]
name = "portfolio-projects"
version = "0.1.0"
description = ""
authors = ["iizhukov <iizhukov@icloud.com>"]
packages = [
    { include = "api" },
    { include = "core" },
    { include = "models" },
    { include = "schemas" },
    { include = "services" },
    { include = "workers" },
    { include = "scripts" }
]

[tool.poetry.dependencies]
python = "^3.12"
fastapi = "0.119.0"
uvicorn = "0.37.0"
pydantic = "2.12.2"
pydantic-settings = "2.11.0"
sqlalchemy = "2.0.44"
asyncpg = "0.30.0"
alembic = "1.17.0"
confluent-kafka = "2.12.0"
grpcio = ">=1.70.0"
protobuf = ">=6.0.0"
httpx = "0.28.1"

[tool.poetry.dependencies.portfolio-shared]
path = "../../shared/python"
develop = true

[tool.poetry.group.dev.dependencies]
grpcio-tools = ">=1.70.0"
pytest = "8.0.0"
pytest-asyncio = "0.23.0"
black = "24.2.0"
flake8 = "7.0.0"
kafka-stubs = "^2.10.0.1"
mypy-protobuf = "^3.6.0"

[tool.poetry.scripts]
generate-grpc = "scripts.generate_grpc:main"
init-db = "scripts.init_db:main"
service = "scripts.run_service:main"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

